Added Pagination, Sorting and Query in Get All Stock

+++ server/Controllers/StockController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using server.Data;
 using server.Dtos.Stock;
+using server.helpers;
 using server.Interfaces;
 using server.Mappers;
 
@@ -17,16 +18,17 @@ namespace server.Controllers
     {
         private readonly IStockRepository _stockRepo = stockRepo;
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            var stocks = await _stockRepo.GetAllAsync();
+            var stocks = await _stockRepo.GetAllAsync(query);
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stock = await _stockRepo.GetAsync(id);
 
             if (stock == null)
@@ -40,22 +42,25 @@ namespace server.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
-            var stockModel = stockDto.ToStockFromCreateDto();
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+            var stockModel = stockDto.ToStockFromCreate();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel);
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto UpdateStockRequestDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stockDto = await _stockRepo.UpdateAsync(id, UpdateStockRequestDto);
             if (stockDto == null) return NotFound();
             return Ok(stockDto.ToStockDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null) return NotFound();
             return NoContent();
